Fix a big source of malloc churn, the outputCreated array.
All of these are arrays of the same size, so there is no point not
just inlining the whole array into the struct.

With this, this can be passed along to subfunctions without copying, saving a _lot_ of malloc churn.

This also involved tightening up const correctness, so we can be assured we are not modifying the
original array when we didn't mean to, which is presumably why the original arrays were being
copied so much. Copying contents of outputCreated arrays is now done via a `memcpy` into the
target struct directly.

Also converts it to a `bool` array, so we can be assured it is 1 or 0 exactly and thus
be able to add them to an index without worrying about non-0 values.