Remove some shutdown checks for performance
handleSelectAndRandom() is called VERY often. Because of this, I understand it seemed appealing to check for shutdown during this function call.
However, we find that the only purpose in checking for shutdown is to write to PB.txt and upload the record to the server.
We've concluded that we will write to PB.txt immediately upon local fastest roadmap discovery, and keep only the shutdown check at the beginning of an iteration.
This way, when prompted to shut down, the program can still respond to this quickly and attempt an upload to the server.
In the event that the user quits out before this happens, then the record will still be submitted on next run.