Fix frame cost bug in tryTossInventoryItem
This function tries to toss all of the first 10 items in our inventory to make room for a recipe output.
Every time stateOK determines this is a valid move, the frame cost is added to the useDescription.
If multiple items can be tossed, then the frame cost is incorrectly summed on top of the previously added frame cost for tossing a different index.
To solve this, we just need to use a copy of useDescription each time we restart the for loop.