Remove critical sec and unnecessary read when writing to PB.txt
We have moved writing to PB.txt to when a new roadmap is discovered.
This occurs inside the "optimize" critical sec, so we can get rid of the pb critical sec.
Also because of this, we no longer need to check if another thread has modified PB.txt,
so we can get rid of reading PB.txt on write attempt.