Get rid of tempFrames function argument
tempFrames was an int ptr passed into various functions related to frame cost calculation.
This variable is unnecessary, as we are always ALSO passing in a MoveDescription,
which can store the associated frame cost. So instead of tempFrames, we just always make sure
to update MoveDescription with the associated frame cost of performing a move.